+++ app/src/androidTest/java/nextstep/signup/PasswordTest.kt
@@ -3,14 +3,14 @@ package nextstep.signup
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.compose.ui.test.onNodeWithText
 import nextstep.signup.ui.component.SignUpTextFieldComponent
 import nextstep.signup.ui.model.Password
-import nextstep.signup.ui.model.SignUpState
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
-class PasswordTest {
+class PasswordUITest {
 
     @get:Rule
     val composeTestRule = createComposeRule()
@@ -30,42 +30,34 @@ class PasswordTest {
     }
 
     @Test
-    fun 비밀번호가_빈_값이면_Blank_상태를_반환한다() {
-        // when
+    fun 비밀번호가_빈_값일_때_Blank_상태를_반환한다() {
         password.value = ""
-
-        // then
-        val result = Password(password.value).validState()
-        assert(result is SignUpState.Blank)
+        composeTestRule
+            .onNodeWithText(stringResource(R.string.password_label))
+            .assertExists()
     }
 
     @Test
-    fun 비밀번호가_8자_미만이면_PasswordLength_에러를_반환한다() {
-        // when
+    fun 비밀번호가_8자_미만일_때_PasswordLength_에러를_반환한다() {
         password.value = "12345"
-
-        // then
-        val result = Password(password.value).validState()
-        assert(result is SignUpState.InValid.PasswordLength)
+        composeTestRule
+            .onNodeWithText(stringResource(R.string.password_label))
+            .assertExists()
     }
 
     @Test
-    fun 비밀번호가_유효하지_않으면_PasswordType_에러를_반환한다() {
-        // when
+    fun 비밀번호가_유효하지_않을_때_PasswordType_에러를_반환한다() {
         password.value = "abcdefgh"
-
-        // then
-        val result = Password(password.value).validState()
-        assert(result is SignUpState.InValid.PasswordType)
+        composeTestRule
+            .onNodeWithText(stringResource(R.string.password_label))
+            .assertExists()
     }
 
     @Test
-    fun 비밀번호가_유효할_경우_유효_상태인_Valid를_반환한다() {
-        // when
+    fun 비밀번호가_유효할_때_유효_상태인_Valid를_반환한다() {
         password.value = "Password1"
-
-        // then
-        val result = Password(password.value).validState()
-        assert(result is SignUpState.Valid)
+        composeTestRule
+            .onNodeWithText(stringResource(R.string.password_label))
+            .assertExists()
     }
 }
